Move functor that detect non void attributes and point in CGAL namespace (before they were in internal); add a missing void typedef.
