+++ Combinatorial_map/include/CGAL/internal/Combinatorial_map_internal_functors.h
@@ -25,7 +25,6 @@
 #include <CGAL/Dimension.h>
 #include <CGAL/Kernel_traits.h>
 #include <vector>
-#include <boost/mpl/has_xxx.hpp>
 
 /* Definition of functors used internally to manage attributes (we need
  * functors as attributes are stored in tuple, thus all the access must be
@@ -65,12 +64,6 @@
  * internal::Test_is_same_attribute_functor<Map1, Map2> to test if two
  *   i-attributes of two darts are isomorphic.
  *
- * internal::Is_attribute_has_non_void_info<Attr> to test if the attribute
- *   Attr is non Void and has an non void Info as inner type
- *
- * internal::Is_attribute_has_point<Attr> to test if the attribute
- *   Attr is non Void and has a Point inner type
- *
  * internal::Reverse_orientation_of_map_functor<CMap> to reverse the
  *   orientation of a whole combinatorial map
  *
@@ -83,6 +76,11 @@
 
 namespace CGAL
 {
+//-----------------------------------------------------------------------------
+template<typename Attr>
+struct Is_attribute_has_non_void_info;
+template<typename Attr>
+struct Is_attribute_has_point;
 // ****************************************************************************
 namespace internal
 {
@@ -637,34 +635,6 @@ struct Is_same_attribute_point_functor<Map1, Map2, T1, T2, false, false, i>
                   typename Map2::Dart_const_handle)
   { return true; }
 };
-// ****************************************************************************
-BOOST_MPL_HAS_XXX_TRAIT_NAMED_DEF(Has_point,Point,false)
-
-template<typename Attr, typename Info=typename Attr::Info>
-struct Is_nonvoid_attribute_has_non_void_info
-{
-  static const bool value=true;
-};
-template<typename Attr>
-struct Is_nonvoid_attribute_has_non_void_info<Attr, void>
-{
-  static const bool value=false;
-};
-
-template<typename Attr>
-struct Is_attribute_has_non_void_info
-{
-  static const bool value=Is_nonvoid_attribute_has_non_void_info<Attr>::value;
-};
-template<>
-struct Is_attribute_has_non_void_info<CGAL::Void>
-{
-  static const bool value=false;
-};
-// ****************************************************************************
-template<typename Attr>
-struct Is_attribute_has_point
-{ static const bool value=Has_point<Attr>::value; };
 // ****************************************************************************
 /// Test if the two darts are associated with the same attribute.
 template<typename Map1, typename Map2>
