+++ Linear_cell_complex/include/CGAL/Cell_attribute_with_point.h
@@ -129,6 +129,7 @@ namespace CGAL {
                            Functor_on_merge_, Functor_on_split_> Base1;
     typedef Point_for_cell<typename LCC::Point> Base2;
 
+    typedef void                            Info;
     typedef typename LCC::Point             Point;
     typedef typename LCC::Dart_handle       Dart_handle;
     typedef typename LCC::Dart_const_handle Dart_const_handle;
